save方法，统一返回新增id

+++ hotdog-saas-tob/hotdog-saas-domain/src/main/java/com/hotdog/saas/domain/repository/EducationCourseClassPersonRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface EducationCourseClassPersonRepository {
 
-    Integer save(EducationCourseClassPerson educationCourseClassPerson);
+    Long save(EducationCourseClassPerson educationCourseClassPerson);
 
     PageResponse<List<EducationCourseClassPerson>> listPage(EducationCourseClassPerson educationCourseClassPerson, PageRequest pageRequest);
