+++ hotdog-saas-tob/hotdog-saas-domain/src/main/java/com/hotdog/saas/domain/repository/LoginLogRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface LoginLogRepository {
 
-    Integer save(LoginLog loginLog);
+    Long save(LoginLog loginLog);
 
     PageResponse<List<LoginLog>> listPage(LoginLog loginLog, PageRequest pageRequest);
 }
